Use the same model for all use cases
